clean up: factor out NonTrivialInliningGraphDecoder. Other refactoring
