+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/ProgressReporter.java
@@ -160,7 +160,6 @@ public class ProgressReporter {
         UNIVERSE("Building universe"),
         PARSING("Parsing methods", true, true),
         INLINING("Inlining methods", true, false),
-        NON_TRIVIAL_INLINING("NonTrivial Inlining", true, false),
         COMPILING("Compiling methods", true, true),
         LAYING_OUT("Laying out methods", true, true),
         CREATING("Creating image", true, true);
@@ -578,10 +577,6 @@ public class ProgressReporter {
 
     public ReporterClosable printInlining() {
         return print(TimerCollection.Registry.INLINE, BuildStage.INLINING);
-    }
-
-    public ReporterClosable printNonTrivialInlining() {
-        return print(TimerCollection.Registry.INLINE, BuildStage.NON_TRIVIAL_INLINING);
     }
 
     public ReporterClosable printCompiling() {
